pytorch/train/classify/requirements.txt
annotated-doc==0.0.4
anyio==4.12.1
certifi==2026.2.25
click==8.3.1
colorama==0.4.6
cuda-bindings==12.9.4
cuda-pathfinder==1.3.4
filelock==3.23.0
flatbuffers==25.12.19
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.7.1
idna==3.11
jinja2==3.1.6
lmdb==1.8.1
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
ml-dtypes==0.5.4
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
onnx==1.20.1
onnxruntime==1.24.1
onnxsim==0.4.36
onnxslim==0.1.87
opencv-python==4.13.0.92
packaging==26.0
pillow==12.1.1
protobuf==6.33.5
pygments==2.19.2
pyyaml==6.0.3
rich==14.3.2
safetensors==0.7.0
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
sympy==1.14.0
timm==1.0.25
torch==2.10.0
torchao==0.16.0
torchvision==0.25.0
tqdm==4.67.3
triton==3.6.0
typer==0.24.1
typing-extensions==4.15.0
